+++ src/app/home/WithdrawRepay/WithdrawRepay.module.css
@@ -102,3 +102,15 @@
   color: var(--secondary-periwinkle);
   font-weight: 700;
 }
+
+.warning {
+  background-color: var(--failed-red);
+  color: var(--failed-red-text);
+  padding: 0.6rem 1.05rem;
+  margin: 1rem 0;
+  border-radius: 12px;
+  display: flex;
+  align-items: center;
+  gap: 0.8rem;
+  font-weight: 500;
+}
