+++ src/app/home/WithdrawRepay/WithdrawRepay.tsx
@@ -14,6 +14,10 @@ import { Quantity } from "ao-tokens";
 import { tokenInput } from "liquidops";
 import { useGetPosition } from "@/hooks/LiquidOpsData/useGetPosition";
 import { useLoadingScreen } from "@/components/LoadingScreen/useLoadingScreen";
+import { useValueLimit } from "@/hooks/data/useValueLimit";
+import { useProtocolStats } from "@/hooks/LiquidOpsData/useProtocolStats";
+import { AnimatePresence, motion } from "framer-motion";
+import { warningVariants } from "@/components/DropDown/FramerMotion";
 
 interface WithdrawRepayProps {
   mode: "withdraw" | "repay";
@@ -63,6 +67,10 @@ const WithdrawRepay: React.FC<WithdrawRepayProps> = ({
       mode === "withdraw" ? unlendError : repayError,
       resetInput,
     );
+
+  const { data: protocolStats, isLoading: isLoadingProtocolStats } =
+    useProtocolStats(ticker.toUpperCase());
+  const [valueLimit, valueLimitReached] = useValueLimit(inputValue, protocolStats);
 
   const calculateMaxAmount = () => {
     if (isLoadingCurrentBalance || !currentBalance)
@@ -166,6 +174,26 @@ const WithdrawRepay: React.FC<WithdrawRepayProps> = ({
       <div className={styles.infoRow}>
         <span className={styles.infoLabel}>Network fee: {networkFee} AO</span>
       </div>
+
+      <AnimatePresence>
+        {valueLimitReached && mode === "withdraw" && (
+          <motion.p
+            variants={warningVariants}
+            initial="hidden"
+            animate="visible"
+            exit="hidden"
+            className={styles.warning}
+          >
+            <Image
+              src="/icons/activity/warning.svg"
+              height={45}
+              width={45}
+              alt="Error icon"
+            />
+            You can only {mode + " "} up to {valueLimit.toLocaleString(undefined, { maximumFractionDigits: 2 }) + " " + ticker}.
+          </motion.p>
+        )}
+      </AnimatePresence>
 
       <SubmitButton
         onSubmit={handleSubmit}
